[Endpoints] Reorganize HostPort code a bit
